Fixes linux build constraints
Added scanner_stub to accomodate for non-linux build
