+++ src/go/guestagent/pkg/procnet/scanner_stub.go
@@ -1,0 +1,34 @@
+//go:build !linux
+
+/*
+Copyright © 2024 SUSE LLC
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package procnet
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/rancher-sandbox/rancher-desktop/src/go/guestagent/pkg/tracker"
+)
+
+type ProcNetScanner struct{}
+
+func NewProcNetScanner(ctx context.Context, tracker tracker.Tracker, scanInterval time.Duration) (*ProcNetScanner, error) {
+	panic("only implemented for Linux")
+}
+
+func (p *ProcNetScanner) ForwardPorts() error {
+	return fmt.Errorf("only implemented for Linux")
+}
